move widgets to screens where appr
